requirements.txt
streamlit
torch
torchvision
timm
opencv-python
pillow
numpy